+++ bioimageio/spec/model/v0_4/schema.py
@@ -1,5 +1,4 @@
 import typing
-import warnings
 from copy import deepcopy
 
 from marshmallow import INCLUDE, RAISE, ValidationError, missing as missing_, pre_load, validates, validates_schema
@@ -10,17 +9,16 @@ from bioimageio.spec.model.v0_3.schema import (
     OnnxWeightsEntry,
     Postprocessing,
     Preprocessing,
-    PytorchScriptWeightsEntry,
     TensorflowJsWeightsEntry,
     TensorflowSavedModelBundleWeightsEntry,
     _WeightsEntryBase,
     _common_sha256_hint,
 )
 from bioimageio.spec.rdf import v0_2 as rdf
+from bioimageio.spec.rdf.v0_2.schema import Author
 from bioimageio.spec.shared import LICENSES, field_validators, fields
 from bioimageio.spec.shared.common import get_args, get_args_flat
 from bioimageio.spec.shared.schema import ImplicitOutputShape, ParametrizedInputShape, SharedBioImageIOSchema
-from bioimageio.spec.shared.utils import is_valid_orcid_id
 from . import raw_nodes
 
 CiteEntry = rdf.schema.CiteEntry
@@ -38,32 +36,6 @@ class Attachments(_BioImageIOSchema):
         fields.Union([fields.URI(), fields.RelativeLocalPath()]),
         bioimageio_description="File attachments; included when packaging the model.",
     )
-
-
-class _Person(_BioImageIOSchema):
-    name = fields.String(bioimageio_description="Full name.")
-    affiliation = fields.String(bioimageio_description="Affiliation.")
-    email = fields.Email()
-    github_user = fields.String(bioimageio_description="GitHub user name.")  # todo: add validation?
-    orcid = fields.String(
-        validate=[
-            field_validators.Length(19),
-            lambda oid: all(oid[idx] == "-" for idx in [4, 9, 14]),
-            lambda oid: is_valid_orcid_id(oid.replace("-", "")),
-        ],
-        bioimageio_description="[orcid](https://support.orcid.org/hc/en-us/sections/360001495313-What-is-ORCID) id "
-        "in hyphenated groups of 4 digits, e.g. '0000-0001-2345-6789' (and [valid]("
-        "https://support.orcid.org/hc/en-us/articles/360006897674-Structure-of-the-ORCID-Identifier"
-        ") as per ISO 7064 11,2.)",
-    )
-
-
-class Author(_Person):
-    name = fields.String(require=True, bioimageio_description="Full name.")
-
-
-class Maintainer(_Person):
-    github_user = fields.String(require=True, bioimageio_description="GitHub user name.")
 
 
 class _TensorBase(_BioImageIOSchema):
@@ -294,7 +266,6 @@ _optional*_ with an asterisk indicates the field is optional depending on the va
         required=True,
         bioimageio_description=rdf.schema.RDF.authors_bioimageio_description,
     )
-    maintainers = fields.List(fields.Nested(Maintainer()), bioimageio_description="Maintainers of this model.")
 
     badges = missing_  # todo: allow badges for Model (RDF has it)
     cite = fields.List(
